Remove copy handlers. Add integration tests
