+++ src/EA.Weee.RequestHandlers.Tests.Unit/EA.Weee.RequestHandlers.Tests.Unit.csproj
@@ -115,7 +115,6 @@
     <Compile Include="Admin\Aatf\CheckAatfApprovalDateChangeHandlerTests.cs" />
     <Compile Include="Admin\Aatf\GetAatfApprovalDateChangeStatusTests.cs" />
     <Compile Include="Admin\CheckApprovalNumberIsUniqueHandlerTests.cs" />
-    <Compile Include="Admin\CopyAatfHandlerTests.cs" />
     <Compile Include="Admin\Aatf\CheckAatfCanBeDeletedHandlerTests.cs" />
     <Compile Include="Admin\Aatf\DeleteAatfHandlerTests.cs" />
     <Compile Include="Admin\Aatf\GetAatfDeletionStatusTests.cs" />
