+++ src/EA.Weee.RequestHandlers/Admin/AddAatfRequestHandler.cs
@@ -66,7 +66,28 @@
                 panArea = await commonDataAccess.FetchLookup<PanArea>(message.Aatf.PanAreaData.Id);
             }
 
-            var aatf = new Domain.AatfReturn.Aatf(
+            if (message.AatfId == null)
+            {
+                var aatf = new Domain.AatfReturn.Aatf(
+                    message.Aatf.Name,
+                    competentAuthority,
+                    message.Aatf.ApprovalNumber,
+                    Enumeration.FromValue<Domain.AatfReturn.AatfStatus>(message.Aatf.AatfStatus.Value),
+                    organisation,
+                    siteAddress,
+                    Enumeration.FromValue<Domain.AatfReturn.AatfSize>(message.Aatf.Size.Value),
+                    message.Aatf.ApprovalDate.GetValueOrDefault(),
+                    contact,
+                    message.Aatf.FacilityType.ToDomainEnumeration<Domain.AatfReturn.FacilityType>(),
+                    message.Aatf.ComplianceYear,
+                    localArea,
+                    panArea);
+
+                await dataAccess.Add<Domain.AatfReturn.Aatf>(aatf);
+            }
+            else
+            {
+                var aatf = new Domain.AatfReturn.Aatf(
                 message.Aatf.Name,
                 competentAuthority,
                 message.Aatf.ApprovalNumber,
@@ -79,9 +100,10 @@
                 message.Aatf.FacilityType.ToDomainEnumeration<Domain.AatfReturn.FacilityType>(),
                 message.Aatf.ComplianceYear,
                 localArea,
-                panArea);
+                panArea, message.AatfId.Value);
 
-            await dataAccess.Add<Domain.AatfReturn.Aatf>(aatf);
+                await dataAccess.Add<Domain.AatfReturn.Aatf>(aatf);
+            }
 
             return true;
         }
