+++ src/EA.Weee.RequestHandlers/EA.Weee.RequestHandlers.csproj
@@ -88,7 +88,6 @@
     <Compile Include="Admin\Aatf\GetAatfApprovalDateChangeStatus.cs" />
     <Compile Include="Admin\Aatf\IGetAatfApprovalDateChangeStatus.cs" />
     <Compile Include="Admin\CheckApprovalNumberIsUniqueHandler.cs" />
-    <Compile Include="Admin\CopyAatfHandler.cs" />
     <Compile Include="Admin\Aatf\CheckAatfCanBeDeletedHandler.cs" />
     <Compile Include="Admin\Aatf\DeleteAatfHandler.cs" />
     <Compile Include="Admin\AatfReports\GetAatfSubmissionHistoryHandler.cs" />
