--- src/EA.Weee.Requests/Admin/CopyAatf.cs
@@ -1,16 +1,0 @@
-﻿namespace EA.Weee.Requests.Admin
-{
-    using EA.Weee.Core.AatfReturn;
-    using Prsd.Core.Mediator;
-    using System;
-    public class CopyAatf : IRequest<bool>
-    {
-        public AatfData Aatf { get; set; }
-
-        public Guid OrganisationId { get; set; }
-
-        public Guid AatfId { get; set; }
-
-        public AatfContactData AatfContact { get; set; }
-    }
-}
